+++ aksel.nav.no/website/app/(routes)/(designsystemet)/_ui/sidebar/Sidebar.module.css
@@ -34,7 +34,6 @@
 .navListGroupLabel {
   padding-block: var(--ax-space-2);
   padding-inline-start: var(--ax-space-8);
-  color: var(--ax-text-neutral-subtle);
   font-weight: var(--ax-font-weight-bold);
 }
